fix: rename 2  Modal
